I am working on a non-PIN version of MDB This is for users who want to have MDB on a screen all players can see, playing "Pass the keyboard" instead of "Pass the laptop". PIN's are also forgotten, so this is helpful for that group of players.
I'm just modifying the standard version of MDB and setting all
PIN's to 0. All user input prompts asking for a PIN will be set to 0.

I wanted a break so I am committing what I currently have.

P.S. We are scrapping the credit card mod for now. (Not sure if that
ever went public, but if it did, were scrapping it until we have an
almost-perfect software)